refactored files of instrumentation
