+++ packages/notest-monitor/src/instrumenter/instrumenter-utils/statements-instrumenters/expression-instrumenter.ts
@@ -1,6 +1,6 @@
 import {Expression, ExpressionStatement, FunctionDeclaration, SyntaxKind} from "ts-morph"
 import {InstrumentStatementInterface} from "./instrument-statement.interface";
-import {collectorCreator} from "../../../../notest-collector/src/collector-creator";
+import {collectorCreator} from "../../../../../notest-collector/src/collector-creator";
 
 export class ExpressionInstrumenter implements InstrumentStatementInterface {
